backend/pyproject.toml
[tool.poetry]
name = "life-dashboard-backend"
version = "0.1.0"
description = "FastAPI backend for the Life Dashboard"
authors = ["Sean May <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { version = "^2.0.30", extras = ["asyncio"] }
asyncpg = "^0.29.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.0"
tenacity = "^8.2.0"
garminconnect = "^0.2.7"
apscheduler = "^3.10.4"
redis = "^5.0.0"
httpx = "^0.27.0"
python-dateutil = "^2.9.0"
loguru = "^0.7.2"
google-cloud-aiplatform = "^1.56.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.11"
google-genai = "^0.5.0"
cryptography = "^42.0.0"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
isort = "^5.13"
mypy = "^1.8"
ruff = "^0.4.0"
pytest = "^8.3"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
